refactor(chat-runtime): share useStableStringArray between home and agent runtimes
